# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APP-STOCK-LNB API",
        "description": "Description de votre API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/articles/batch": {
            "post": {
                "tags": [
                    "Articles"
                ],
                "summary": "Créer plusieurs articles en lot",
                "operationId": "44f9c655c7d56546beaa1a60fa0f4879",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "articles": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "id_cat",
                                                "libelle",
                                                "code_article",
                                                "stock_alerte"
                                            ],
                                            "properties": {
                                                "id_cat": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "libelle": {
                                                    "type": "string",
                                                    "example": "Chaussures de sport"
                                                },
                                                "code_article": {
                                                    "type": "string",
                                                    "example": "ART-2025-001"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Description optionnelle"
                                                },
                                                "stock_alerte": {
                                                    "type": "integer",
                                                    "example": 5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Articles créés",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostResourceResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/articles/{id}": {
            "put": {
                "tags": [
                    "Articles"
                ],
                "summary": "Mettre à jour un article",
                "operationId": "2717d5a8e3b1f16637aded325347f9ad",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'article",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_cat",
                                    "libelle",
                                    "code_article",
                                    "stock_alerte"
                                ],
                                "properties": {
                                    "id_cat": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "libelle": {
                                        "type": "string",
                                        "example": "Chaussures modifiées"
                                    },
                                    "code_article": {
                                        "type": "string",
                                        "example": "ART-2025-002"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Description mise à jour"
                                    },
                                    "stock_alerte": {
                                        "type": "integer",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Article mis à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostResourceResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Articles"
                ],
                "summary": "Supprimer un article",
                "operationId": "97f34592611a1dd53a086865e29706c9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'article",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article supprimé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Article supprimé avec succès"
                                        },
                                        "data": {
                                            "type": "null",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Article": {
                "title": "Article",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "id_cat": {
                        "type": "integer",
                        "example": 3
                    },
                    "libelle": {
                        "type": "string",
                        "example": "Chaussures de sport"
                    },
                    "code_article": {
                        "type": "string",
                        "example": "ART-2025-001"
                    },
                    "description": {
                        "type": "string",
                        "example": "Chaussures confortables pour la course"
                    },
                    "stock_alerte": {
                        "type": "integer",
                        "example": 5
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-25T14:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-25T14:00:00Z"
                    },
                    "categorie": {
                        "$ref": "#/components/schemas/CategorieArticle"
                    },
                    "stock": {
                        "$ref": "#/components/schemas/Stock"
                    }
                },
                "type": "object"
            },
...
            "CategorieArticle": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 3
                    },
                    "libelle_categorie_article": {
                        "type": "string",
                        "example": "Chaussures"
                    }
                },
                "type": "object"
            },
...
            "Stock": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "id_Article": {
                        "type": "integer",
                        "example": 1
                    },
                    "Qte_actuel": {
                        "type": "integer",
                        "example": 10
                    }
                },
                "type": "object"
            },
...
            "PostResourceResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Opération réussie"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Article"
                        }
                    }
                },
                "type": "object"
            }
        }
    },
...
}